reorganized file structure for namefield component
